REFACT: 라우팅 관련 코드는 routes 폴더에, 페이지 구성은 pages 폴더에
