--- src/components/todo/TodoDetail.tsx
@@ -1,31 +1,0 @@
-import { useEffect, useState } from "react";
-import { useQuery } from "react-query";
-import { useTodo } from "hooks/todo/useTodo";
-import { usePath } from "hooks/.commons/usePath";
-import TodoDetailEdit from "./TodoDetailEdit";
-import TodoDetailItem from "./TodoDetailItem";
-import "assets/css/TodoDetail.css";
-
-const TodoDetail = () => {
-  const [isEditMode, setEditMode] = useState(false);
-  const id = usePath();
-  const { getTodoById } = useTodo();
-  const { data, refetch: refresh } = useQuery("getTodo", () => getTodoById(id));
-  const todo = data?.data;
-
-  useEffect(() => {
-    refresh();
-  }, [id, refresh]);
-
-  return isEditMode ? (
-    <TodoDetailEdit
-      id={id}
-      originalTodo={data?.data}
-      setEditMode={setEditMode}
-    />
-  ) : (
-    <TodoDetailItem id={id} todo={todo} setEditMode={setEditMode} />
-  );
-};
-
-export default TodoDetail;
