+++ src/pages/TodoCreatePage.tsx
@@ -9,8 +9,9 @@ import Input from "components/.commons/Input";
 import Button from "components/.commons/Button";
 import Textarea from "components/.commons/Textarea";
 import "assets/css/TodoForm.css";
+import TodoWrapper from "components/todo/TodoWrapper";
 
-const TodoForm = () => {
+const TodoCreatePage = () => {
   const [todo, setTodo] = useState({
     title: "",
     content: "",
@@ -38,29 +39,33 @@ const TodoForm = () => {
   };
 
   return (
-    <form className="todo-detail__wrapper todo-form" onSubmit={submitTodo}>
-      <Input
-        className="todo-detail__title"
-        placeholder="할일 제목을 입력해주세요"
-        value={todo.title}
-        onChange={(value: string) => amendState(todo, setTodo, "title", value)}
-      />
-      <Textarea
-        className="todo-detail__content"
-        placeholder="할일 세부내용을 적어주세요"
-        value={todo.content}
-        onChange={(value: string) =>
-          amendState(todo, setTodo, "content", value)
-        }
-      />
-      <Button
-        className="todo-form__submit-button"
-        value="새 할일 추가"
-        type="submit"
-        isFilled
-      />
-    </form>
+    <TodoWrapper>
+      <form className="todo-detail__wrapper todo-form" onSubmit={submitTodo}>
+        <Input
+          className="todo-detail__title"
+          placeholder="할일 제목을 입력해주세요"
+          value={todo.title}
+          onChange={(value: string) =>
+            amendState(todo, setTodo, "title", value)
+          }
+        />
+        <Textarea
+          className="todo-detail__content"
+          placeholder="할일 세부내용을 적어주세요"
+          value={todo.content}
+          onChange={(value: string) =>
+            amendState(todo, setTodo, "content", value)
+          }
+        />
+        <Button
+          className="todo-form__submit-button"
+          value="새 할일 추가"
+          type="submit"
+          isFilled
+        />
+      </form>
+    </TodoWrapper>
   );
 };
 
-export default TodoForm;
+export default TodoCreatePage;
